Move AddNodeMenu into its own tsx file
